Run UMAP embedding for cell-gene graph built up by caclust.

This function takes cacomp and caclust object as input to calculate UMAP embedding of cell-gene graph in several different ways:

'SNNdist'(Default): run UMAP on the distance matrix of cell-gene SNN graph built up by caclust, which is '1-adj(SNN)'.

'spectral': run UMAP on the selected eigenvectors of cell-gene graph laplacian (only eligiable when algorithm is set as 'spectral' in 'caclust' function)

'ca': run UMAP on the singular vectors from Correspondence Analysis.

Usage

run_biMAP(
  obj,
  caobj = NULL,
  k = 30,
  rand_seed = 2358,
  method = "SNNdist",
  use_SNN = TRUE,
  features = NULL
)

Arguments

obj

results from biclustering of class "caclust"

caobj

A cacomp object with principal and standard coordinates calculated. Only needs to be supplied when using method "ca".

k

integer. Number of nearest neighbours to use to compute UMAP.

rand_seed

integer. Random seed for UMAP.

method

Can be either "SNNdist", "spectral" or "ca". When using "ca", a "cacomp" object has to be provided for caobj.

use_SNN

TRUE/FALSE. This parameter only works when method == 'ca'. If TRUE, it will calculate bimap embedding of genes and cells contained in the SNN graph together with the given 'fearures' if any. If FALSE, only calculate bimap embedding of cells contained in the SNN graph together with the given 'fearures' if any.

features

character/vector of gene names. This parameter only works when method == 'ca'. 'ca' method allows visualization of genes which are absent from SNN graph, so the 'features' could be genes detected by our biclustering algorithm, and genes which goes through CA analysis. Setting use_SNN = FALSE allows users to have a visualization of the feature genes user defined/interested.

Value

caclust object with biMAP coordinates stored in the bimap slot.