internal helper to plot scatter biMAP

internal helper to plot scatter biMAP

Usage

scatter_plot(
  umap_cells,
  umap_genes,
  color_by,
  colors,
  color_genes,
  cell_size,
  cell_alpha,
  gene_size,
  gene_alpha,
  interact_genes,
  interact_cells
)

Arguments

umap_cells

biMAP coordinates data.frame subsetted to cells.

umap_genes

biMAP coordinates data.frame subsetted to genes.

color_by

Either "type" or "cluster". "type" colors by the type (cell or gene) while "cluster" colors by the assigned cluster. If meta_df is provided a column from the data.frame can be specified.

colors

named vector of of length equal to the group size.

color_genes

logical. If TRUE colors genes by color_by.

cell_size

integer. Size of the cells in the plot.

cell_alpha

numeric. Alpha value of cells (between 0 and 1).

gene_size

integer. Size of the genes in the plot.

gene_alpha

numeric. Alpha value of genes (between 0 and 1).

interact_genes

character string with text and labels for genes in interactive plots.

interact_cells

character string with text and labels for cells in interactive plot.

Value

ggplot object